all routes with jwt token authentication

+++ src/routes/memories.ts
@@ -28,12 +28,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
-    const { id } = paramsSchema.parse(request.params)
+    const { id } = paramsSchema.parse(req.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
@@ -41,36 +41,44 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
+    const { sub } = req.user
+
+    if (!memory.isPublic && memory.userId !== sub) {
+      return res.status(401).send()
+    }
+
     return memory
   })
 
-  app.post('/memories', async (request) => {
+  app.post('/memories', async (req) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
     })
 
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
+    const { content, coverUrl, isPublic } = bodySchema.parse(req.body)
+
+    const { sub } = req.user
 
     const memory = await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPublic,
-        userId: '5ac8562f-5c67-49da-8319-5646cf8b8b6d',
+        userId: sub,
       },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
-    const { id } = paramsSchema.parse(request.params)
+    const { id } = paramsSchema.parse(req.params)
 
     const bodySchema = z.object({
       content: z.string(),
@@ -78,9 +86,21 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false),
     })
 
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
+    const { content, coverUrl, isPublic } = bodySchema.parse(req.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    const { sub } = req.user
+
+    if (memory.userId !== sub) {
+      return res.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -94,12 +114,23 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
-    const { id } = paramsSchema.parse(request.params)
+    const { id } = paramsSchema.parse(req.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    const { sub } = req.user
+    if (memory.userId !== sub) {
+      return res.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
